+++ rflow/rflow/doctype/ourcompany/ourcompany.json
@@ -3,7 +3,9 @@
  "actions": [],
  "allow_rename": 1,
  "creation": "2022-10-14 15:43:40.784917",
+ "default_view": "List",
  "doctype": "DocType",
+ "document_type": "Setup",
  "editable_grid": 1,
  "engine": "InnoDB",
  "field_order": [
@@ -146,10 +148,11 @@
    "label": "Tax Ammount"
   }
  ],
+ "force_re_route_to_default_view": 1,
  "index_web_pages_for_search": 1,
  "is_submittable": 1,
  "links": [],
- "modified": "2022-12-12 17:57:56.306833",
+ "modified": "2022-12-15 20:53:52.381779",
  "modified_by": "Administrator",
  "module": "Rflow",
  "name": "OurCompany",
@@ -188,6 +191,8 @@
    "write": 1
   }
  ],
+ "search_fields": "director,company_address",
+ "show_name_in_global_search": 1,
  "sort_field": "modified",
  "sort_order": "DESC",
  "states": []
