+++ rflow/rflow/doctype/rieltflow/rieltflow.json
@@ -3,6 +3,7 @@
  "actions": [],
  "allow_rename": 1,
  "creation": "2022-10-07 15:45:45.704219",
+ "default_view": "List",
  "doctype": "DocType",
  "document_type": "Document",
  "editable_grid": 1,
@@ -284,6 +285,7 @@
    "label": "KPP"
   }
  ],
+ "force_re_route_to_default_view": 1,
  "is_calendar_and_gantt": 1,
  "is_submittable": 1,
  "links": [
@@ -294,7 +296,7 @@
   }
  ],
  "max_attachments": 2,
- "modified": "2022-12-14 15:36:14.868528",
+ "modified": "2022-12-15 20:43:28.931300",
  "modified_by": "Administrator",
  "module": "Rflow",
  "name": "RieltFlow",
@@ -321,6 +323,7 @@
    "amend": 1,
    "cancel": 1,
    "create": 1,
+   "delete": 1,
    "email": 1,
    "export": 1,
    "print": 1,
@@ -333,7 +336,8 @@
    "write": 1
   }
  ],
- "search_fields": "company, address",
+ "search_fields": "date,agent",
+ "show_name_in_global_search": 1,
  "sort_field": "modified",
  "sort_order": "DESC",
  "states": [],
