+++ rflow/rflow/doctype/taxcode/taxcode.json
@@ -1,7 +1,9 @@
 {
+ "_liked_by": "[\"Administrator\"]",
  "actions": [],
  "allow_rename": 1,
  "creation": "2022-10-10 21:20:05.183023",
+ "default_view": "List",
  "doctype": "DocType",
  "editable_grid": 1,
  "engine": "InnoDB",
@@ -65,9 +67,10 @@
    "fieldtype": "Section Break"
   }
  ],
+ "force_re_route_to_default_view": 1,
  "index_web_pages_for_search": 1,
  "links": [],
- "modified": "2022-12-12 17:01:31.856367",
+ "modified": "2022-12-15 20:54:43.139095",
  "modified_by": "Administrator",
  "module": "Rflow",
  "name": "TaxCode",
@@ -101,6 +104,7 @@
    "write": 1
   }
  ],
+ "search_fields": "area",
  "sort_field": "modified",
  "sort_order": "DESC",
  "states": []
